refactor(apps): split files over 250 lines into focused sibling modules
Extract mappers, validation, types and seed data into sibling files so no
source file exceeds the 250-line guideline. Behaviour is unchanged: moved
code keeps its logic, and modules that others import re-export the moved
types. Also ignore the local postman folder.
