+++ .gitignore
@@ -10,3 +10,4 @@ coverage/
 .env.*
 !.env.example
 .DS_Store
+postman/
