+++ apps/beneficiary-service/src/beneficiary/beneficiary.service.ts
@@ -1,10 +1,7 @@
 import { addDays } from '@armman/core';
 import {
-  conflict,
-  decryptPii,
   encryptPii,
   hashForSearch,
-  HttpError,
   notFound,
   normalizeForSearch,
   unprocessable,
@@ -14,11 +11,9 @@ import type {
   BeneficiaryStatusHistory,
 } from '../../../../node_modules/.prisma/client-beneficiary-service';
 import type { BeneficiaryStatus, CaseType } from './beneficiary.constants';
-import type {
-  BeneficiaryListFilters,
-  BeneficiaryRepository,
-  DuplicateSearchTokens,
-} from './beneficiary.repository';
+import { buildSearchTokens, evaluateDuplicateMatch } from './beneficiary.duplicate-detection';
+import { computeBmi, withDecryptedName } from './beneficiary.mapper';
+import type { BeneficiaryListFilters, BeneficiaryRepository } from './beneficiary.repository';
 import type { CreateBeneficiaryInput } from './dto/create-beneficiary.dto';
 
 const GESTATION_DAYS = 280;
@@ -34,152 +29,6 @@ export interface ListBeneficiariesQuery {
   /** Raw search text — hashed the same way as duplicate-detection tokens (exact match only). */
   name?: string;
   mobileNumber?: string;
-}
-
-/**
- * The shape `findDuplicateCandidate` returns — the matched BeneficiaryCase
- * with just the fields FR-S-2.4/2.5 need off its current summary. Structural
- * (not the full Prisma type) so the decision logic isn't coupled to it.
- */
-interface DuplicateMatch {
-  id: string;
-  currentStatus: string;
-  currentSummary: {
-    dateOfDelivery: Date | null;
-    closureDate: Date | null;
-    lmpDate: Date | null;
-  } | null;
-}
-
-interface PiiRow {
-  id: string;
-  fullNameEnc: Buffer;
-  villageId: string | null;
-  padaId: string | null;
-  healthSubCentreId: string | null;
-  phcId: string | null;
-  healthBlockId: string | null;
-  dateOfBirth: Date | null;
-  sex: string | null;
-  stateId: string | null;
-  districtId: string | null;
-  talukaId: string | null;
-}
-
-/**
- * Projects a raw beneficiary_case Prisma row down to EXACTLY the fields the
- * API documents (beneficiaryCaseSchema / beneficiaryCaseDetailSchema in
- * beneficiary.controller.ts), decrypting `pii.fullName` for display. Every
- * level is allow-listed — the case, `pii`, and each nested relation — so
- * internal columns (createdByUserId/updatedByUserId/isDeleted/deletedAt,
- * encrypted/hash PII columns, undocumented case fields like
- * pregnancySequenceNo/journeyEndDate, and nested audit columns) can never
- * leak into a response even as the Prisma rows gain columns.
- *
- * Nested relations are only projected when present, so this serves both the
- * list rows (case + pii only) and the detail view (case + pii + mother/child
- * details + consent + risk/status).
- */
-function withDecryptedName<T extends { pii: PiiRow; [k: string]: unknown }>(caseRow: T) {
-  const c = caseRow as Record<string, unknown>;
-  const pii = caseRow.pii;
-
-  const projected: Record<string, unknown> = {
-    id: c.id,
-    localCaseUuid: c.localCaseUuid,
-    piiId: c.piiId,
-    projectId: c.projectId,
-    sakhiId: c.sakhiId,
-    caseType: c.caseType,
-    registrationDate: c.registrationDate,
-    currentStatus: c.currentStatus,
-    currentPhase: c.currentPhase,
-    beneficiaryTypeLookupId: c.beneficiaryTypeLookupId,
-    caseTypeLookupId: c.caseTypeLookupId,
-    previousBeneficiaryId: c.previousBeneficiaryId,
-    motherBeneficiaryId: c.motherBeneficiaryId,
-    journeyStartDate: c.journeyStartDate,
-    createdAt: c.createdAt,
-    updatedAt: c.updatedAt,
-    pii: {
-      id: pii.id,
-      fullName: decryptPii(pii.fullNameEnc),
-      villageId: pii.villageId,
-      padaId: pii.padaId,
-      healthSubCentreId: pii.healthSubCentreId,
-      phcId: pii.phcId,
-      healthBlockId: pii.healthBlockId,
-      dateOfBirth: pii.dateOfBirth,
-      sex: pii.sex,
-      stateId: pii.stateId,
-      districtId: pii.districtId,
-      talukaId: pii.talukaId,
-    },
-  };
-
-  // Nested relations — only projected when the query included them (detail view).
-  const mother = c.motherCaseDetails as Record<string, unknown> | null | undefined;
-  if (mother !== undefined) {
-    projected.motherCaseDetails = mother
-      ? {
-          lmpDate: mother.lmpDate,
-          eddDate: mother.eddDate,
-          gravida: mother.gravida,
-          parity: mother.parity,
-          heightCm: mother.heightCm,
-          bmiAtRegistration: mother.bmiAtRegistration,
-        }
-      : null;
-  }
-  const child = c.childCaseDetails as Record<string, unknown> | null | undefined;
-  if (child !== undefined) {
-    projected.childCaseDetails = child
-      ? {
-          motherBeneficiaryId: child.motherBeneficiaryId,
-          dateOfBirth: child.dateOfBirth,
-          sex: child.sex,
-          birthWeightKg: child.birthWeightKg,
-          birthLengthCm: child.birthLengthCm,
-          prematureFlag: child.prematureFlag,
-          linkedAncCase: child.linkedAncCase,
-        }
-      : null;
-  }
-  const consents = c.consentRecords as Record<string, unknown>[] | undefined;
-  if (consents !== undefined) {
-    projected.consentRecords = consents.map((r) => ({
-      consentType: r.consentType,
-      consentStatus: r.consentStatus,
-      consentDate: r.consentDate,
-      capturedByUserId: r.capturedByUserId,
-    }));
-  }
-  const risks = c.riskConditionSummaries as Record<string, unknown>[] | undefined;
-  if (risks !== undefined) {
-    projected.riskConditionSummaries = risks.map((r) => ({
-      riskConditionId: r.riskConditionId,
-      phase: r.phase,
-      latestGrade: r.latestGrade,
-      latestAssessedAt: r.latestAssessedAt,
-      everHighestGrade: r.everHighestGrade,
-      everAtRiskFlag: r.everAtRiskFlag,
-      currentReferralTriggerFlag: r.currentReferralTriggerFlag,
-      currentHrVisitTriggerFlag: r.currentHrVisitTriggerFlag,
-    }));
-  }
-  const history = c.statusHistory as Record<string, unknown>[] | undefined;
-  if (history !== undefined) {
-    projected.statusHistory = history.map((h) => ({
-      fromStatus: h.fromStatus,
-      toStatus: h.toStatus,
-      reasonCode: h.reasonCode,
-      changedByUserId: h.changedByUserId,
-      changedAt: h.changedAt,
-      notes: h.notes,
-    }));
-  }
-
-  return projected;
 }
 
 /** Business logic for the beneficiary enrollment lifecycle. */
@@ -233,12 +82,12 @@ export class BeneficiaryService {
       throw unprocessable('Consent not received. Registration cannot proceed.');
     }
 
-    const searchTokens = this.buildSearchTokens(dto);
+    const searchTokens = buildSearchTokens(dto);
 
     if (!dto.acknowledgeDuplicate) {
       const match = await this.repository.findDuplicateCandidate(searchTokens);
       if (match) {
-        this.evaluateDuplicateMatch(match, dto);
+        evaluateDuplicateMatch(match, dto);
       }
     }
 
@@ -320,86 +169,4 @@ export class BeneficiaryService {
       statusHistory: [] as BeneficiaryStatusHistory[],
     };
   }
-
-  private buildSearchTokens(dto: CreateBeneficiaryInput): DuplicateSearchTokens {
-    const geographyParts = [dto.pii.villageId, dto.pii.padaId].filter(Boolean).join('|');
-    const dob = dto.case.caseType === 'CHILD' ? dto.childDetails?.dateOfBirth : dto.pii.dateOfBirth;
-
-    return {
-      nameToken: hashForSearch(normalizeForSearch(dto.pii.fullName)),
-      caseTypeLookupId: dto.case.caseTypeLookupId,
-      dobToken: dob ? hashForSearch(dob.toISOString().slice(0, 10)).toString('base64') : null,
-      phoneHash: dto.pii.phone ? hashForSearch(normalizeForSearch(dto.pii.phone)) : null,
-      geographyToken: geographyParts
-        ? hashForSearch(normalizeForSearch(geographyParts)).toString('base64')
-        : null,
-      lmpDateToken: dto.motherDetails
-        ? hashForSearch(dto.motherDetails.lmpDate.toISOString().slice(0, 10)).toString('base64')
-        : null,
-    };
-  }
-
-  /**
-   * Decides how to handle a duplicate-detection match, per SRS FR-S-2.4/2.5.
-   * Throws to block or prompt; returns normally to allow the enrollment.
-   *
-   * The matched case's `currentSummary` carries delivery/closure/status/LMP.
-   *
-   * - FR-S-2.5 (re-enrolment): matched case is JOURNEY_COMPLETE/CLOSED, has a
-   *   confirmed delivery, and the new LMP differs from the matched case's LMP
-   *   → surface the specific "is this a new pregnancy?" prompt (409 with a
-   *   RE_ENROLLMENT reason in details) so the client can confirm and resubmit
-   *   with acknowledgeDuplicate.
-   * - FR-S-2.4 (new pregnancy): matched case has BOTH a delivery date AND a
-   *   closure date → treat as a completed prior journey, allow the new
-   *   enrollment (return normally).
-   * - FR-S-2.4 (hard duplicate): anything else — including a matched case
-   *   with no summary row at all (SRS-literal: "if both not present →
-   *   duplicate, registration cannot proceed") → block with a 409.
-   */
-  private evaluateDuplicateMatch(match: DuplicateMatch, dto: CreateBeneficiaryInput): void {
-    const summary = match.currentSummary;
-    const hasDelivery = Boolean(summary?.dateOfDelivery);
-    const hasClosure = Boolean(summary?.closureDate);
-    const isCompletedJourney =
-      match.currentStatus === 'JOURNEY_COMPLETE' || match.currentStatus === 'CLOSED';
-
-    const newLmp = dto.motherDetails?.lmpDate;
-    const priorLmp = summary?.lmpDate;
-    const lmpDiffers =
-      newLmp != null &&
-      priorLmp != null &&
-      newLmp.toISOString().slice(0, 10) !== priorLmp.toISOString().slice(0, 10);
-
-    // FR-S-2.5 — re-enrolment for a new pregnancy after a completed journey.
-    if (isCompletedJourney && hasDelivery && lmpDiffers) {
-      throw new HttpError(
-        409,
-        'A previous record exists for this beneficiary. Is this a new pregnancy?',
-        {
-          reason: 'RE_ENROLLMENT',
-          existingBeneficiaryId: match.id,
-          resolution: 'Resubmit with acknowledgeDuplicate: true to enroll a new pregnancy.',
-        },
-      );
-    }
-
-    // FR-S-2.4 — both delivery AND closure exist → completed prior journey,
-    // this is a legitimate new pregnancy; allow it through.
-    if (hasDelivery && hasClosure) {
-      return;
-    }
-
-    // FR-S-2.4 — otherwise a hard duplicate (including no summary row at all).
-    throw conflict(
-      `A possible duplicate beneficiary already exists (beneficiaryId: ${match.id}). ` +
-        'Resubmit with acknowledgeDuplicate: true to proceed anyway.',
-    );
-  }
-}
-
-function computeBmi(heightCm: number | undefined, weightKg: number | undefined): number | null {
-  if (!heightCm || !weightKg) return null;
-  const heightM = heightCm / 100;
-  return Math.round((weightKg / (heightM * heightM)) * 100) / 100;
 }
